Avoid loading advice classes in agent class loader, part 2
