+++ instrumentation/jaxrs/jaxrs-2.0/jaxrs-2.0-jersey-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jaxrs/v2_0/JerseyRequestContextInstrumentation.java
@@ -26,7 +26,7 @@ import net.bytebuddy.asm.Advice.Local;
 public class JerseyRequestContextInstrumentation extends AbstractRequestContextInstrumentation {
   @Override
   protected String abortAdviceName() {
-    return ContainerRequestContextAdvice.class.getName();
+    return getClass().getName() + "$ContainerRequestContextAdvice";
   }
 
   public static class ContainerRequestContextAdvice {
