+++ instrumentation/jaxrs/jaxrs-2.0/jaxrs-2.0-resteasy-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jaxrs/v2_0/Resteasy30RequestContextInstrumentation.java
@@ -27,7 +27,7 @@ import org.jboss.resteasy.core.interception.PostMatchContainerRequestContext;
 public class Resteasy30RequestContextInstrumentation extends AbstractRequestContextInstrumentation {
   @Override
   protected String abortAdviceName() {
-    return ContainerRequestContextAdvice.class.getName();
+    return getClass().getName() + "$ContainerRequestContextAdvice";
   }
 
   public static class ContainerRequestContextAdvice {
