+++ instrumentation/jaxrs/jaxrs-2.0/jaxrs-2.0-resteasy-3.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jaxrs/v2_0/Resteasy31RequestContextInstrumentation.java
@@ -27,7 +27,7 @@ import org.jboss.resteasy.core.interception.jaxrs.PostMatchContainerRequestConte
 public class Resteasy31RequestContextInstrumentation extends AbstractRequestContextInstrumentation {
   @Override
   protected String abortAdviceName() {
-    return ContainerRequestContextAdvice.class.getName();
+    return getClass().getName() + "$ContainerRequestContextAdvice";
   }
 
   public static class ContainerRequestContextAdvice {
