+++ instrumentation/spring/spring-webflux-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/webflux/server/DispatcherHandlerInstrumentation.java
@@ -5,16 +5,23 @@
 
 package io.opentelemetry.javaagent.instrumentation.spring.webflux.server;
 
+import static io.opentelemetry.javaagent.instrumentation.spring.webflux.server.SpringWebfluxHttpServerTracer.tracer;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
+import io.opentelemetry.api.trace.SpanKind;
+import io.opentelemetry.context.Context;
+import io.opentelemetry.context.Scope;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
+import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
+import org.springframework.web.server.ServerWebExchange;
+import reactor.core.publisher.Mono;
 
 public class DispatcherHandlerInstrumentation implements TypeInstrumentation {
 
@@ -31,6 +38,43 @@ public class DispatcherHandlerInstrumentation implements TypeInstrumentation {
             .and(named("handle"))
             .and(takesArgument(0, named("org.springframework.web.server.ServerWebExchange")))
             .and(takesArguments(1)),
-        DispatcherHandlerAdvice.class.getName());
+        this.getClass().getName() + "$HandleAdvice");
+  }
+
+  /**
+   * This is 'top level' advice for Webflux instrumentation. This handles creating and finishing
+   * Webflux span.
+   */
+  public static class HandleAdvice {
+
+    @Advice.OnMethodEnter(suppress = Throwable.class)
+    public static void methodEnter(
+        @Advice.Argument(0) ServerWebExchange exchange,
+        @Advice.Local("otelScope") Scope otelScope,
+        @Advice.Local("otelContext") Context otelContext) {
+
+      otelContext = tracer().startSpan("DispatcherHandler.handle", SpanKind.INTERNAL);
+      // Unfortunately Netty EventLoop is not instrumented well enough to attribute all work to the
+      // right things so we have to store the context in request itself.
+      exchange.getAttributes().put(AdviceUtils.CONTEXT_ATTRIBUTE, otelContext);
+
+      otelScope = otelContext.makeCurrent();
+    }
+
+    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
+    public static void methodExit(
+        @Advice.Thrown Throwable throwable,
+        @Advice.Argument(0) ServerWebExchange exchange,
+        @Advice.Return(readOnly = false) Mono<Void> mono,
+        @Advice.Local("otelScope") Scope otelScope,
+        @Advice.Local("otelContext") Context otelContext) {
+      if (throwable == null && mono != null) {
+        mono = AdviceUtils.setPublisherSpan(mono, otelContext);
+      } else if (throwable != null) {
+        AdviceUtils.finishSpanIfPresent(exchange, throwable);
+      }
+      otelScope.close();
+      // span finished in SpanFinishingSubscriber
+    }
   }
 }
