+++ instrumentation/spring/spring-webflux-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/webflux/server/RouterFunctionInstrumentation.java
@@ -17,8 +17,13 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
+import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
+import org.springframework.web.reactive.function.server.HandlerFunction;
+import org.springframework.web.reactive.function.server.RouterFunction;
+import org.springframework.web.reactive.function.server.ServerRequest;
+import reactor.core.publisher.Mono;
 
 public class RouterFunctionInstrumentation implements TypeInstrumentation {
 
@@ -47,6 +52,26 @@ public class RouterFunctionInstrumentation implements TypeInstrumentation {
                 takesArgument(
                     0, named("org.springframework.web.reactive.function.server.ServerRequest")))
             .and(takesArguments(1)),
-        RouterFunctionAdvice.class.getName());
+        this.getClass().getName() + "$Route");
+  }
+
+  /**
+   * This advice is responsible for setting additional span parameters for routes implemented with
+   * functional interface.
+   */
+  public static class RouteAdvice {
+
+    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
+    public static void methodExit(
+        @Advice.This RouterFunction thiz,
+        @Advice.Argument(0) ServerRequest serverRequest,
+        @Advice.Return(readOnly = false) Mono<HandlerFunction<?>> result,
+        @Advice.Thrown Throwable throwable) {
+      if (throwable == null) {
+        result = result.doOnSuccessOrError(new RouteOnSuccessOrError(thiz, serverRequest));
+      } else {
+        AdviceUtils.finishSpanIfPresent(serverRequest, throwable);
+      }
+    }
   }
 }
